Rename PhpMyAdmin\Dbi to PhpMyAdmin\Dbal
Rename Dbi (Database Interface) to Dbal (Database Abstraction Layer).
